Distinguish between modification and transformation
TinySpline now distinguishes between modifications and
transformations. Modifications directly modify a spline and
are less complex. Transformations, on the other hand,
create a copy of the spline to modify, apply modifications,
and return the copy. In general, transformations are more
complex. Furthermore, transformations are supposed to be
chained (e.g. spline.derive().toBeziers...).